fix(rust): audit round 2 — crash recovery, panic paths and integer edges
Round 2 used two more reviewers (panic/boundary correctness, end-to-end crash
recovery). Fourteen findings; this PR takes all of them.

Crash recovery (the reviewer drove the real appender with kill -9):
- heap fallback: the cache file was read into the region but never written
  back or cleared, so a drained cache was appended again on every later start
  — once per start, forever, behind a legitimate-looking "begin of mmap"
  banner. Cleared after the drain and on close when there is no mapping.
  (This was a regression from the mmap-fallback change in the previous PR.)
- sparse cache file: set_len() alone leaves a hole, and the first store into
  the mapping is what allocates — a full disk turns that into SIGBUS. The
  backing store is pre-allocated by writing zeros, and a failure falls back to
  the heap path, exactly like mars/comm/mmap_util.cc.
- open(): the drained leftover was dropped when the writer thread could not be
  spawned. It is now written through a synchronous drain on that path.
- category: the "we opened the appender" flag was lost when another thread
  won the registration race, so nothing ever closed it.
- async wake-ups are coalesced (sync_channel(1) + try_send) instead of
  queueing one message per write; Close still blocks so it is never dropped.

Panic paths:
- a non-UTF-8 log_dir became "" and the open failed, silently disabling
  logging for the whole process. Paths are now read as bytes.
- eprintln! panics when stderr cannot be written; on the async thread there is
  no barrier, so that panic ended logging permanently. The error is ignored.
- the per-thread recursion counter leaked on any panic, leaving that thread
  permanently above MAX_RECURSION. It is now restored by a Drop guard.
- file_util's split index used `index + 1` (wraps in release, panics in debug
  on the async thread) and the timespan helpers multiplied without
  saturating. Both fixed.
- LogBuffer could emit a record with no tailer byte when is_compress == false
  filled the region; the region now reserves it.
- ffi: the output slice was sized from the caller's claimed length rather
  than the bytes actually written; pid was truncated through i32.
- the header documented a "Mars" default for name_prefix that the code
  deliberately does not implement.

+++ rust/crates/mars-xlog-ffi/src/state.rs
@@ -43,7 +43,7 @@ pub fn level_enabled(level: i32) -> bool {
 
 /// `xlogger_pid()` — the OS process id.
 pub fn pid() -> i64 {
-    i64::from(std::process::id() as i32)
+    std::process::id() as i64
 }
 
 /// `xlogger_tid()` — the OS thread id, so records written through the FFI can
